@@ -1,80 +1,67 @@
 {
-  "name": "linkifyjs",
-  "version": "2.1.9",
+  "name": "linkify",
+  "version": "3.0.0",
   "description": "Intelligent link recognition, made easy",
   "repository": {
     "type": "git",
-    "url": "https://github.com/SoapBox/linkifyjs.git"
+    "url": "git+https://github.com/Soapbox/linkifyjs.git"
   },
   "main": "index.js",
   "scripts": {
-    "test": "nyc --reporter=html --reporter=text gulp test",
-    "test:ci": "gulp test-ci",
-    "coverage": "nyc report --reporter=text-lcov | coveralls"
+    "clean": "rm -rf dist",
+    "coverage": "nyc report --reporter=text-lcov | coveralls",
+    "build": "npm run build --workspaces",
+    "build:ci": "for dir in `ls packages` ; do cd packages/${dir} && npm run build && cd ../.. ; done",
+    "copy": "copyfiles -f packages/*/dist/*.js dist",
+    "dist": "run-s clean build copy",
+    "dist:ci": "run-s clean build:ci copy",
+    "lint": "eslint *.js packages/**/*.js test/**/*.js",
+    "test": "nyc --reporter=lcov --reporter=text mocha test/index.js",
+    "test:ci": "karma start test/ci.conf.js --single-run"
   },
-  "author": "SoapBox Innovations (@SoapBoxHQ)",
+
+  "author": "Soapbox Innovations (@SoapBoxHQ)",
   "license": "MIT",
-  "keywords": [
-    "jquery",
-    "link",
-    "autolink",
-    "text",
-    "url",
-    "email",
-    "hashtag",
-    "hashtags",
-    "mention",
-    "mentions"
-  ],
   "devDependencies": {
-    "babel-core": "^6.26.3",
-    "babel-plugin-add-module-exports": "^0.2.1",
-    "babel-plugin-transform-es2015-modules-amd": "^6.24.1",
-    "babel-polyfill": "^6.26.0",
-    "babel-preset-es2015": "^6.24.1",
-    "chai": "^4.2.0",
-    "coveralls": "^3.0.9",
+    "@babel/core": "^7.13.10",
+    "@babel/eslint-parser": "^7.13.10",
+    "@babel/eslint-plugin": "^7.13.10",
+    "@babel/preset-env": "^7.13.10",
+    "@babel/register": "^7.13.8",
+    "@nfrasser/simple-html-tokenizer": "^0.5.11-1",
+    "@rollup/plugin-babel": "^5.3.0",
+    "@rollup/plugin-commonjs": "^17.1.0",
+    "@rollup/plugin-node-resolve": "^11.2.0",
+    "chai": "^4.3.4",
+    "copyfiles": "^2.4.1",
+    "coveralls": "^3.1.0",
+    "eslint": "^7.22.0",
+    "eslint-plugin-mocha": "^8.1.0",
     "glob": "^7.1.6",
-    "gulp": "^4.0.2",
-    "gulp-babel": "^7.0.1",
-    "gulp-clean": "^0.3.2",
-    "gulp-cli": "^2.2.0",
-    "gulp-concat": "^2.6.0",
-    "gulp-eslint": "^4.0.2",
-    "gulp-mocha": "^5.0.0",
-    "gulp-rename": "^1.4.0",
-    "gulp-replace": "^0.5.4",
-    "gulp-uglify": "^1.5.4",
-    "gulp-util": "^3.0.7",
-    "gulp-wrap": "^0.13.0",
-    "jquery": ">= 1.11.0",
+    "jquery": "^3.6.0",
     "jsdom": "^11.12.0",
-    "karma": "^4.4.1",
-    "karma-chrome-launcher": "^2.2.0",
-    "karma-firefox-launcher": "^1.3.0",
-    "karma-phantomjs-launcher": "^1.0.4",
-    "karma-qunit": "^1.0.0",
-    "karma-sauce-launcher": "^2.0.2",
-    "lazypipe": "^1.0.2",
-    "lodash": "^4.17.15",
-    "merge-stream": "^1.0.0",
-    "mocha": "^5.2.0",
-    "nyc": "^13.3.0",
-    "phantomjs-prebuilt": "^2.1.15",
-    "qunitjs": "^1.14.0",
-    "react": "^16.12.0",
-    "react-dom": "^16.12.0",
-    "requirejs": "^2.3.6",
-    "rollup": "~0.40.0",
-    "simple-html-tokenizer": "git+https://github.com/nfrasser/simple-html-tokenizer.git#04799f4638ec5ed903a4e5aa6e832269fa59be6b",
-    "through2": "^2.0.5"
+    "karma": "6.2.0",
+    "karma-browserstack-launcher": "^1.6.0",
+    "karma-chrome-launcher": "^3.1.0",
+    "karma-firefox-launcher": "^2.1.0",
+    "karma-qunit": "^4.1.2",
+    "mocha": "^8.3.2",
+    "mocha-lcov-reporter": "^1.3.0",
+    "npm-run-all": "^4.1.5",
+    "nyc": "^15.1.0",
+    "qunit": "^2.14.1",
+    "react": "^17.0.1",
+    "react-dom": "^17.0.1",
+    "request": "^2.88.2",
+    "rollup": "^2.42.1",
+    "rollup-plugin-terser": "^7.0.2",
+    "typescript": "^4.4.2"
   },
-  "peerDependencies": {
-    "jquery": ">= 1.11.0",
-    "react": ">= 0.14.0",
-    "react-dom": ">= 0.14.0"
-  },
+  "private": true,
   "engines": {
     "node": ">=8"
-  }
+  },
+  "workspaces": [
+    "./packages/*"
+  ]
 }
