@@ -1,17 +1,15 @@
 // Karma Chrome configuration
 // Just opens Google Chrome for testing
 
-var
-base = require('./conf'),
-extend = require('lodash/extend');
+const base = require('./conf');
 
 module.exports = function (config) {
+	config.set({
+		...base,
 
-  config.set(extend(base, {
-
-    // level of logging
-    // possible values: config.LOG_DISABLE || config.LOG_ERROR || config.LOG_WARN || config.LOG_INFO || config.LOG_DEBUG
-    logLevel: config.LOG_INFO,
-    browsers: ['Firefox']
-  }));
+		// level of logging
+		// possible values: config.LOG_DISABLE || config.LOG_ERROR || config.LOG_WARN || config.LOG_INFO || config.LOG_DEBUG
+		logLevel: config.LOG_INFO,
+		browsers: ['Firefox']
+	});
 };
